Use 4 digit octal mode
This work only updates whitespace and converts any numeric octal values
to strings.  The is to accommodate the warning below, that is showing
itself in recent version of Puppet.

Warning: Non-string values for the file mode property are deprecated. It
must be a string, either a symbolic mode like 'o+w,a+r' or an octal
representation like '0644' or '755'.