Support more key types in ssh::hosts
Currently, only sshdsa keys are exported through the ssh::hosts class.
The key is also exported blindly without checking for the fact's
presence.  This is a problem because a number of other host keys are
potentially missing from the network of host keys.  This work extends
the ssh::hosts class to include more key types for export which will
also increase the number of keys collected from the network.