Fix key hostnames
Without this change, keys are distributed, but the titlevar is used to
represent the hostname of the system exporting the key.  This is a
problem because it means that the keys are not validated correctly
against the local cache in ssh_known_hosts.

This work adds an aliases variable with trusted defaults to allow the
hostname to be verified as an alias to the keyname.  Allows for the
distribution of multiple key types under the same set of host aliases.