Use an integers when generating ssh keys
This work migrates all references of SSH key length to be integers.
Also here we fix an issue where the default target would contain a key
that doesn't match the key type.  Now we default to am empty value and
interpret the default based on the key type rather than requiring the
user to specify both values.