Allow managing ssh_config(5) configuration file
This is heavily based on the code to manage the `sshd_config(5)`
configuration file.  For consistency, it does not enforce types for the
configuration parameters.

Some basic testing has been setup, but since it implies a lot of
duplication, ssh::client::host and ssh::client::match defined types are
not tested for now.