Exclude the .git directory from built package
When building a package using the Puppet Development Kit (default
packaging system when building a package with Puppet >= 6.0.0), the .git
directory is not automatically excluded, resulting in a package larger
than necessary and which is longer to process by r10k.

Add a .pdkignore file listing the .git directory as being ignored when
building the package to fix these issues.