Improve ssh_known_hosts content generation
This module rely on the `sshkeys` type which is now provided by the
`puppetlabs/sshkeys_core` module (since Puppet 6.0 which moved a bunch of
modules out of puppet core).

In order to avoid duplicate resources, this module used custom resource
names in the form "sshdsakey-<hostname>", "sshecdsakey-<hostname>" and
"sshed25519key-<hostname>", which are present in the managed
`ssh_known_hosts` file, and can be read in some situations like a shell
completion.  However, such names being just workarounds, they are
unlikely to resolve to something valid, and if they do, they may resolve
to the wrong target.

Version 2.0.0 of the `sshkeys_core` module [reworked] the way resource
title is managed, allowing us to have a clean `ssh_known_hosts` file
generated.  All versions of Puppet 7 ship with a compatible version of
`sshkeys_core` bundled, yet Puppet 6 does not.  However, it is possible
to use a newer `sshkeys_core` on Puppet 6 by adding it to the Puppetfile
as any other module.

To avoid usage of an incompatible module, explicitely add the dependency
to the metadata file.

[reworked]: https://github.com/puppetlabs/puppetlabs-sshkeys_core/pull/27